Move control flow demos to tests
