--- icn-ccl/test_enhanced_features.rs
@@ -1,177 +1,0 @@
-#![allow(clippy::uninlined_format_args)]
-
-use icn_ccl::compile_ccl_source_to_wasm;
-
-fn main() {
-    println!("🌟 ICN CCL Enhanced Features Test 🌟\n");
-
-    // Test 1: Advanced Functions with Parameters
-    println!("=== Test 1: Advanced Function Parameters ===");
-    let function_test = r#"
-        fn process_data(base: Integer, multiplier: Integer, label: String) -> Integer {
-            let result = base * multiplier;
-            return result;
-        }
-        
-        fn calculate_bonus(score: Integer, level: Integer) -> Integer {
-            let bonus = score + level * 10;
-            return bonus;
-        }
-        
-        fn run() -> Integer {
-            let total = process_data(5, 3, "test");
-            let bonus = calculate_bonus(total, 2);
-            return bonus;
-        }
-    "#;
-
-    match compile_ccl_source_to_wasm(function_test) {
-        Ok((wasm_bytes, metadata)) => {
-            println!("✅ Advanced functions compiled successfully!");
-            println!("📦 WASM size: {} bytes", wasm_bytes.len());
-            println!("📋 Exports: {:?}", metadata.exports);
-        }
-        Err(e) => {
-            println!("❌ Advanced functions failed: {:?}", e);
-        }
-    }
-
-    // Test 2: String Operations
-    println!("\n=== Test 2: String Operations ===");
-    let string_test = r#"
-        fn build_message(prefix: String, suffix: String) -> String {
-            let message = prefix + suffix;
-            return message;
-        }
-        
-        fn run() -> String {
-            return build_message("Hello ", "ICN!");
-        }
-    "#;
-
-    match compile_ccl_source_to_wasm(string_test) {
-        Ok((wasm_bytes, metadata)) => {
-            println!("✅ String operations compiled successfully!");
-            println!("📦 WASM size: {} bytes", wasm_bytes.len());
-            println!("📋 Exports: {:?}", metadata.exports);
-        }
-        Err(e) => {
-            println!("❌ String operations failed: {:?}", e);
-        }
-    }
-
-    // Test 3: Array Handling (will partially work - arrays are simplified)
-    println!("\n=== Test 3: Array Literals ===");
-    let array_test = r#"
-        fn process_scores(scores: Array<Integer>) -> Integer {
-            return 100; // Simplified - real implementation would process array
-        }
-        
-        fn run() -> Integer {
-            let numbers = [1, 2, 3, 4, 5];
-            return process_scores(numbers);
-        }
-    "#;
-
-    match compile_ccl_source_to_wasm(array_test) {
-        Ok((wasm_bytes, metadata)) => {
-            println!("✅ Array literals compiled successfully!");
-            println!("📦 WASM size: {} bytes", wasm_bytes.len());
-            println!("📋 Exports: {:?}", metadata.exports);
-        }
-        Err(e) => {
-            println!("❌ Array literals failed: {:?}", e);
-        }
-    }
-
-    // Test 4: Complex Governance Contract
-    println!("\n=== Test 4: Governance Contract Example ===");
-    let governance_test = r#"
-        fn calculate_voting_power(mana: Mana, reputation: Integer) -> Integer {
-            let base_power = mana / 100;
-            let bonus = reputation * 2;
-            return base_power + bonus;
-        }
-        
-        fn check_quorum(total_votes: Integer, min_threshold: Integer) -> Bool {
-            return total_votes >= min_threshold;
-        }
-        
-        fn validate_proposal(submitter_mana: Mana, min_mana: Mana) -> Bool {
-            return submitter_mana >= min_mana;
-        }
-        
-        fn run() -> Integer {
-            let voter_mana = 1000;
-            let voter_reputation = 50;
-            let voting_power = calculate_voting_power(voter_mana, voter_reputation);
-            
-            if check_quorum(voting_power, 25) {
-                return voting_power;
-            } else {
-                return 0;
-            }
-        }
-    "#;
-
-    match compile_ccl_source_to_wasm(governance_test) {
-        Ok((wasm_bytes, metadata)) => {
-            println!("✅ Governance contract compiled successfully!");
-            println!("📦 WASM size: {} bytes", wasm_bytes.len());
-            println!("📋 Exports: {:?}", metadata.exports);
-            println!("🏛️ Functions: calculate_voting_power, check_quorum, validate_proposal, run");
-        }
-        Err(e) => {
-            println!("❌ Governance contract failed: {:?}", e);
-        }
-    }
-
-    // Test 5: Economic Policy Contract
-    println!("\n=== Test 5: Economic Policy Contract ===");
-    let economic_test = r#"
-        fn calculate_mana_cost(cpu_cores: Integer, memory_mb: Integer, duration_ms: Integer) -> Mana {
-            let cpu_cost = cpu_cores * 10;
-            let memory_cost = memory_mb / 256;
-            let time_cost = duration_ms / 1000;
-            return cpu_cost + memory_cost + time_cost;
-        }
-        
-        fn apply_reputation_discount(base_cost: Mana, reputation: Integer) -> Mana {
-            if reputation > 80 {
-                return base_cost * 80 / 100; // 20% discount
-            } else {
-                return base_cost;
-            }
-        }
-        
-        fn run() -> Mana {
-            let base_cost = calculate_mana_cost(4, 1024, 30000);
-            let final_cost = apply_reputation_discount(base_cost, 85);
-            return final_cost;
-        }
-    "#;
-
-    match compile_ccl_source_to_wasm(economic_test) {
-        Ok((wasm_bytes, metadata)) => {
-            println!("✅ Economic policy compiled successfully!");
-            println!("📦 WASM size: {} bytes", wasm_bytes.len());
-            println!("📋 Exports: {:?}", metadata.exports);
-            println!("💰 Functions: calculate_mana_cost, apply_reputation_discount, run");
-        }
-        Err(e) => {
-            println!("❌ Economic policy failed: {:?}", e);
-        }
-    }
-
-    println!("\n🎉 CCL Enhanced Features Test Complete!");
-    println!("🚀 New Capabilities Added:");
-    println!("   • ✅ Function parameters with multiple types");
-    println!("   • ✅ String type and concatenation operations");
-    println!("   • ✅ Array literals and type checking");
-    println!("   • ✅ Mana, Did, Proposal, Vote governance types");
-    println!("   • ✅ Complex conditional logic (if statements)");
-    println!("   • ✅ Local variable scoping and shadowing");
-    println!("   • ✅ Multi-function composition");
-    println!("   • ✅ Type compatibility checking");
-    println!("\n💡 Ready for real governance and economic policies!");
-}
